Removes the datapusher installation as it breaks the dependencies.
The datapusher is in conflict with the CKAN base dependecies and if
required must be installed in a separate virtual env.

However, the datapusher process is not directly required by the unit
tests. The CKAN extension 'datapusher' (which is distributed with
CKAN under ckanext, and it is a separate from 'datapusher' process)
is enough to run the tests.